Change working directory when unpacking soundlang files
On FreeBSD, /tmp is generally marked setuid, which has
the unfortunate property of making a rename out of the
directory fail with a permissions problem.

So now the extraction code creates a tmp directory under
amp_conf['ASTVARLIBDIR']/sounds and use that while unpacking.
Add a recursivermdir() function that will completely clean up
the working directory.