FREEPBX-23827: Fix problem delay when loading the list of available languages in the packages page
Modify the way in which the list of languages is loaded to avoid the delay that occurs each time the web is loaded.
Now when loading the packages page only the installed languages are shown.
A button is added so that the list also shows the languages available for installation from the freepbx website.
Now the table is updated via ajax.
Now savesettings is executed via ajax.
Add box with wait message, while running long processes in the background via ajax.
Now the names of the tables are obtained from a private variable $tables of the class.
Add button to download the list of available languages from the freepbx website and prevent it from running whenever the list of packages is loaded.
Create variables within the class with default values such as language and formats.
get language info package via ajax.
Replace ajax.php by var.
Fix help msg in the console command.
Update i18n
Show process installing.
When saving settings, show the update progress window of the language packs.
Remove external execution of "TAR" for decompression, php's own function is used to extract the contents of the tar file